Waveform data removed from this repository
Previously, this repository included a collection of waveform files,
which were very large (and also should not be shared publicly due to
the rules of the 2015 PhysioNet Challenge.)

The 'main' branch was rewritten by running the following command:

   git filter-branch -f --tree-filter 'if [ -d record-files ]; then find record-files \( -name \*.alm -o -name \*.cba -o -name \*.hea -o -name \*.mat \) -delete; fi' HEAD

References to historical commits may therefore be inaccurate.

Example data will be re-added once we can find some that is small and
unproblematic.